新增 18.09.8 mips64 构建相关 patch

+++ components/engine/daemon/daemon_unix.go
@@ -174,8 +174,8 @@ func getBlkioWeightDevices(config containertypes.Resources) ([]specs.LinuxWeight
 		}
 		weight := weightDevice.Weight
 		d := specs.LinuxWeightDevice{Weight: &weight}
-		d.Major = int64(unix.Major(stat.Rdev))
-		d.Minor = int64(unix.Minor(stat.Rdev))
+		d.Major = int64(unix.Major(uint64(stat.Rdev)))
+		d.Minor = int64(unix.Minor(uint64(stat.Rdev)))
 		blkioWeightDevices = append(blkioWeightDevices, d)
 	}
 
@@ -245,8 +245,8 @@ func getBlkioThrottleDevices(devs []*blkiodev.ThrottleDevice) ([]specs.LinuxThro
 			return nil, err
 		}
 		d := specs.LinuxThrottleDevice{Rate: d.Rate}
-		d.Major = int64(unix.Major(stat.Rdev))
-		d.Minor = int64(unix.Minor(stat.Rdev))
+		d.Major = int64(unix.Major(uint64(stat.Rdev)))
+		d.Minor = int64(unix.Minor(uint64(stat.Rdev)))
 		throttleDevices = append(throttleDevices, d)
 	}
